be/pyproject.toml
[project]
name = "deriv-hackathon"
version = "0.1.0"
description = "AI-Powered Payments Approval & Fraud Intelligence"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.3",
    "uvicorn[standard]>=0.40.0",
    "langchain>=1.0",
    "langchain-google-genai>=4.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.2.1",
    "langchain-community>=0.4.1",
    "sqlalchemy>=2.0.46",
    "asyncpg>=0.31.0",
    "alembic>=1.18.3",
    "psycopg2-binary>=2.9.11",
    "langchain-chroma>=0.1.2",
    "chromadb>=0.5.0",
    "hdbscan>=0.8.33",
    "scikit-learn>=1.3.0",
    "langchain-tavily>=0.1.0",
]
